console: Move api key form to console components
