+++ pkg/webui/console/components/api-key-form/create.js
@@ -16,15 +16,15 @@ import React from 'react'
 import bind from 'autobind-decorator'
 
 import ApiKeyModal from '../api-key-modal'
-import PropTypes from '../../lib/prop-types'
-import sharedMessages from '../../lib/shared-messages'
-import SubmitBar from '../submit-bar'
-import FormField from '../form/field'
-import FormSubmit from '../form/submit'
-import SubmitButton from '../submit-button'
-import Input from '../input'
-import Message from '../../lib/components/message'
-import RightsGroup from '../../console/components/rights-group'
+import PropTypes from '../../../lib/prop-types'
+import sharedMessages from '../../../lib/shared-messages'
+import SubmitBar from '../../../components/submit-bar'
+import FormField from '../../../components/form/field'
+import FormSubmit from '../../../components/form/submit'
+import SubmitButton from '../../../components/submit-button'
+import Input from '../../../components/input'
+import Message from '../../../lib/components/message'
+import RightsGroup from '../../../console/components/rights-group'
 import ApiKeyForm from './form'
 import validationSchema from './validation-schema'
