+++ pkg/webui/console/components/api-key-form/edit.js
@@ -16,18 +16,18 @@ import React from 'react'
 import bind from 'autobind-decorator'
 import { defineMessages } from 'react-intl'
 
-import diff from '../../lib/diff'
-import PropTypes from '../../lib/prop-types'
-import sharedMessages from '../../lib/shared-messages'
-import SubmitBar from '../submit-bar'
-import ModalButton from '../button/modal-button'
-import toast from '../toast'
-import Message from '../../lib/components/message'
-import FormField from '../form/field'
-import FormSubmit from '../form/submit'
-import SubmitButton from '../submit-button'
-import Input from '../input'
-import RightsGroup from '../../console/components/rights-group'
+import diff from '../../../lib/diff'
+import PropTypes from '../../../lib/prop-types'
+import sharedMessages from '../../../lib/shared-messages'
+import SubmitBar from '../../../components/submit-bar'
+import ModalButton from '../../../components/button/modal-button'
+import toast from '../../../components/toast'
+import Message from '../../../lib/components/message'
+import FormField from '../../../components/form/field'
+import FormSubmit from '../../../components/form/submit'
+import SubmitButton from '../../../components/submit-button'
+import Input from '../../../components/input'
+import RightsGroup from '../../../console/components/rights-group'
 import ApiKeyForm from './form'
 import validationSchema from './validation-schema'
