+++ pkg/webui/console/views/gateway-api-key-edit/index.js
@@ -24,7 +24,7 @@ import sharedMessages from '../../../lib/shared-messages'
 import Spinner from '../../../components/spinner'
 import Message from '../../../lib/components/message'
 import IntlHelmet from '../../../lib/components/intl-helmet'
-import { ApiKeyEditForm } from '../../../components/api-key-form'
+import { ApiKeyEditForm } from '../../components/api-key-form'
 
 import { getGatewayApiKey } from '../../store/actions/gateway'
 import { getGatewaysRightsList } from '../../store/actions/gateways'
