+++ pkg/webui/locales/en.json
@@ -1,13 +1,4 @@
 {
-  "components.api-key-form.edit.deleteKey": "Delete Key",
-  "components.api-key-form.edit.modalWarning": "Are you sure you want to delete the {keyName} API Key? Deleting an API Key cannot be undone!",
-  "components.api-key-form.edit.updateSuccess": "Successfully updated API Key",
-  "components.api-key-form.edit.deleteSuccess": "Successfully deleted API Key",
-  "components.api-key-modal.index.title": "Please copy newly created API Key",
-  "components.api-key-modal.index.subtitle": "You won't be able to view the key afterward",
-  "components.api-key-modal.index.buttonMessage": "I have copied the key",
-  "components.api-key-modal.index.grantedRights": "Granted Rights",
-  "components.api-key-modal.index.description": "Your API Key has been created successfully.\nNote: After closing this window, the value of the key secret will not be accessible anymore.\nMake sure to copy and store it in a safe place now.",
   "components.event.index.eventData": "Event Data",
   "components.events.index.truncated": "Events have been truncated",
   "components.events.index.showing": "Showing all available events",
@@ -27,6 +18,16 @@
   "components.safe-inspector.index.copyClipboard": "Copy to clipboard",
   "components.safe-inspector.index.arrayFormatting": "Toggle array formatting",
   "components.safe-inspector.index.byteOrder": "Switch byte order",
+  "console.components.api-key-form.edit.deleteKey": "Delete Key",
+  "console.components.api-key-form.edit.modalWarning": "Are you sure you want to delete the {keyName} API Key? Deleting an API Key cannot be undone!",
+  "console.components.api-key-form.edit.updateSuccess": "Successfully updated API Key",
+  "console.components.api-key-form.edit.deleteSuccess": "Successfully deleted API Key",
+  "console.components.api-key-modal.index.title": "Please copy newly created API Key",
+  "console.components.api-key-modal.index.subtitle": "You won't be able to view the key afterward",
+  "console.components.api-key-modal.index.buttonMessage": "I have copied the key",
+  "console.components.api-key-modal.index.grantedRights": "Granted Rights",
+  "console.components.api-key-modal.index.description": "Your API Key has been created successfully.\nNote: After closing this window, the value of the key secret will not be accessible anymore.\nMake sure to copy and store it in a safe place now.",
+  "console.components.collaborator-form.index.cannotModifyRightAll": "This user possesses universal admin rights that cannot be modified.",
   "console.components.gateway-data-form.index.enforced": "Enforced",
   "console.components.gateway-data-form.index.dutyCycle": "Duty Cycle",
   "console.components.gateway-data-form.index.gatewayIdPlaceholder": "my-new-gateway",
@@ -99,9 +100,6 @@
   "console.views.application-add.index.appNamePlaceholder": "My New Application",
   "console.views.application-add.index.appDescPlaceholder": "Description for my new application",
   "console.views.application-add.index.createApplication": "Create Application",
-  "console.views.application-collaborator-edit.index.deleteSuccess": "Successfully removed collaborator",
-  "console.views.application-collaborator-edit.index.updateSuccess": "Successfully updated collaborator rights",
-  "console.views.application-collaborator-edit.index.modalWarning": "Are you sure you want to remove {collaboratorId} as a collaborator?",
   "console.views.application-data.index.appData": "Application Data",
   "console.views.application-general-settings.index.basics": "Basics",
   "console.views.application-general-settings.index.deleteApp": "Delete application",
@@ -167,7 +165,6 @@
   "lib.shared-messages.add": "Add",
   "lib.shared-messages.addApiKey": "Add API Key",
   "lib.shared-messages.addApplication": "Add Application",
-  "lib.shared-messages.addCollaborator": "Add Collaborator",
   "lib.shared-messages.addDevice": "Add Device",
   "lib.shared-messages.addGateway": "Add Gateway",
   "lib.shared-messages.addWebhook": "Add Webhook",
@@ -187,10 +184,15 @@
   "lib.shared-messages.brand": "Brand",
   "lib.shared-messages.cancel": "Cancel",
   "lib.shared-messages.clear": "Clear",
+  "lib.shared-messages.collaboratorAdd": "Add Collaborator",
+  "lib.shared-messages.collaboratorDeleteSuccess": "Successfully removed collaborator",
   "lib.shared-messages.collaboratorEdit": "Edit {collaboratorId}",
   "lib.shared-messages.collaboratorEditRights": "Edit rights of {collaboratorId}",
   "lib.shared-messages.collaboratorId": "Collaborator ID",
+  "lib.shared-messages.collaboratorModalWarning": "Are you sure you want to remove {collaboratorId} as a collaborator?",
+  "lib.shared-messages.collaboratorRemove": "Collaborator Remove",
   "lib.shared-messages.collaborators": "Collaborators",
+  "lib.shared-messages.collaboratorUpdateSuccess": "Successfully updated collaborator rights",
   "lib.shared-messages.connecting": "Connecting",
   "lib.shared-messages.createApiKey": "Create API Key",
   "lib.shared-messages.created": "Created",
