+++ pkg/webui/locales/xx.json
@@ -1,13 +1,4 @@
 {
-  "components.api-key-form.edit.deleteKey": "Xxxxxx Xxx",
-  "components.api-key-form.edit.modalWarning": "Xxx xxx xxxx xxx xxxx xx xxxxxx xxx {keyName} XXX Xxx? Xxxxxxxx xx XXX Xxx xxxxxx xx xxxxxx!",
-  "components.api-key-form.edit.updateSuccess": "Xxxxxxxxxxxx xxxxxxx XXX Xxx",
-  "components.api-key-form.edit.deleteSuccess": "Xxxxxxxxxxxx xxxxxxx XXX Xxx",
-  "components.api-key-modal.index.title": "Xxxxxx xxxx xxxxx xxxxxxx XXX Xxx",
-  "components.api-key-modal.index.subtitle": "Xxx xxx'x xx xxxx xx xxxx xxx xxx xxxxxxxxx",
-  "components.api-key-modal.index.buttonMessage": "X xxxx xxxxxx xxx xxx",
-  "components.api-key-modal.index.grantedRights": "Xxxxxxx Xxxxxx",
-  "components.api-key-modal.index.description": "Xxxx XXX Xxx xxx xxxx xxxxxxx xxxxxxxxxxxx.\nXxxx: Xxxxx xxxxxxx xxxx xxxxxx, xxx xxxxx xx xxx xxx xxxxxx xxxx xxx xx xxxxxxxxxx xxxxxxx.\nXxxx xxxx xx xxxx xxx xxxxx xx xx x xxxx xxxxx xxx.",
   "components.event.index.eventData": "Xxxxx Xxxx",
   "components.events.index.truncated": "Xxxxxx xxxx xxxx xxxxxxxxx",
   "components.events.index.showing": "Xxxxxxx xxx xxxxxxxxx xxxxxx",
@@ -27,6 +18,16 @@
   "components.safe-inspector.index.copyClipboard": "Xxxx xx xxxxxxxxx",
   "components.safe-inspector.index.arrayFormatting": "Xxxxxx xxxxx xxxxxxxxxx",
   "components.safe-inspector.index.byteOrder": "Xxxxxx xxxx xxxxx",
+  "console.components.api-key-form.edit.deleteKey": "Xxxxxx Xxx",
+  "console.components.api-key-form.edit.modalWarning": "Xxx xxx xxxx xxx xxxx xx xxxxxx xxx {keyName} XXX Xxx? Xxxxxxxx xx XXX Xxx xxxxxx xx xxxxxx!",
+  "console.components.api-key-form.edit.updateSuccess": "Xxxxxxxxxxxx xxxxxxx XXX Xxx",
+  "console.components.api-key-form.edit.deleteSuccess": "Xxxxxxxxxxxx xxxxxxx XXX Xxx",
+  "console.components.api-key-modal.index.title": "Xxxxxx xxxx xxxxx xxxxxxx XXX Xxx",
+  "console.components.api-key-modal.index.subtitle": "Xxx xxx'x xx xxxx xx xxxx xxx xxx xxxxxxxxx",
+  "console.components.api-key-modal.index.buttonMessage": "X xxxx xxxxxx xxx xxx",
+  "console.components.api-key-modal.index.grantedRights": "Xxxxxxx Xxxxxx",
+  "console.components.api-key-modal.index.description": "Xxxx XXX Xxx xxx xxxx xxxxxxx xxxxxxxxxxxx.\nXxxx: Xxxxx xxxxxxx xxxx xxxxxx, xxx xxxxx xx xxx xxx xxxxxx xxxx xxx xx xxxxxxxxxx xxxxxxx.\nXxxx xxxx xx xxxx xxx xxxxx xx xx x xxxx xxxxx xxx.",
+  "console.components.collaborator-form.index.cannotModifyRightAll": "Xxxx xxxx xxxxxxxxx xxxxxxxxx xxxxx xxxxxx xxxx xxxxxx xx xxxxxxxx.",
   "console.components.gateway-data-form.index.enforced": "Xxxxxxxx",
   "console.components.gateway-data-form.index.dutyCycle": "Xxxx Xxxxx",
   "console.components.gateway-data-form.index.gatewayIdPlaceholder": "xx-xxx-xxxxxxx",
@@ -99,9 +100,6 @@
   "console.views.application-add.index.appNamePlaceholder": "Xx Xxx Xxxxxxxxxxx",
   "console.views.application-add.index.appDescPlaceholder": "Xxxxxxxxxxx xxx xx xxx xxxxxxxxxxx",
   "console.views.application-add.index.createApplication": "Xxxxxx Xxxxxxxxxxx",
-  "console.views.application-collaborator-edit.index.deleteSuccess": "Xxxxxxxxxxxx xxxxxxx xxxxxxxxxxxx",
-  "console.views.application-collaborator-edit.index.updateSuccess": "Xxxxxxxxxxxx xxxxxxx xxxxxxxxxxxx xxxxxx",
-  "console.views.application-collaborator-edit.index.modalWarning": "Xxx xxx xxxx xxx xxxx xx xxxxxx {collaboratorId} xx x xxxxxxxxxxxx?",
   "console.views.application-data.index.appData": "Xxxxxxxxxxx Xxxx",
   "console.views.application-general-settings.index.basics": "Xxxxxx",
   "console.views.application-general-settings.index.deleteApp": "Xxxxxx xxxxxxxxxxx",
@@ -167,7 +165,6 @@
   "lib.shared-messages.add": "Xxx",
   "lib.shared-messages.addApiKey": "Xxx XXX Xxx",
   "lib.shared-messages.addApplication": "Xxx Xxxxxxxxxxx",
-  "lib.shared-messages.addCollaborator": "Xxx Xxxxxxxxxxxx",
   "lib.shared-messages.addDevice": "Xxx Xxxxxx",
   "lib.shared-messages.addGateway": "Xxx Xxxxxxx",
   "lib.shared-messages.addWebhook": "Xxx Xxxxxxx",
@@ -187,10 +184,15 @@
   "lib.shared-messages.brand": "Xxxxx",
   "lib.shared-messages.cancel": "Xxxxxx",
   "lib.shared-messages.clear": "Xxxxx",
+  "lib.shared-messages.collaboratorAdd": "Xxx Xxxxxxxxxxxx",
+  "lib.shared-messages.collaboratorDeleteSuccess": "Xxxxxxxxxxxx xxxxxxx xxxxxxxxxxxx",
   "lib.shared-messages.collaboratorEdit": "Xxxx {collaboratorId}",
   "lib.shared-messages.collaboratorEditRights": "Xxxx xxxxxx xx {collaboratorId}",
   "lib.shared-messages.collaboratorId": "Xxxxxxxxxxxx XX",
+  "lib.shared-messages.collaboratorModalWarning": "Xxx xxx xxxx xxx xxxx xx xxxxxx {collaboratorId} xx x xxxxxxxxxxxx?",
+  "lib.shared-messages.collaboratorRemove": "Xxxxxxxxxxxx Xxxxxx",
   "lib.shared-messages.collaborators": "Xxxxxxxxxxxxx",
+  "lib.shared-messages.collaboratorUpdateSuccess": "Xxxxxxxxxxxx xxxxxxx xxxxxxxxxxxx xxxxxx",
   "lib.shared-messages.connecting": "Xxxxxxxxxx",
   "lib.shared-messages.createApiKey": "Xxxxxx XXX Xxx",
   "lib.shared-messages.created": "Xxxxxxx",
